+++ QualtricsStudyGroupSurveys/survey_generator.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta
 from dotenv import load_dotenv
 from QualtricsStudyGroupSurveys import QualtricsConnection
 
-from questions import (
+from .questions import (
     meet_count_question,
     no_meeting_explanation,
     no_meeting_upload_question,
@@ -14,7 +14,7 @@ from questions import (
     meeting_duration_question,
     who_did_you_not_meet_with_question
 )
-from blocks import create_block_and_return_id, get_loop_options
+from .blocks import create_block_and_return_id, get_loop_options
 
 
 load_dotenv()
@@ -70,8 +70,6 @@ def generate_survey(qualtrics_connection: QualtricsConnection, survey_id, date_c
     
     who_did_you_not_meet_block_id = create_block_and_return_id(qualtrics_connection, survey_id, "Who didn't you meet with?")
     qualtrics_connection.add_question(survey_id, who_did_you_not_meet_block_id, who_did_you_not_meet_with_question(people))
-
-    print("Added all questions")
