+++ main.py
@@ -1,8 +1,11 @@
 import json
 import sys
-from QualtricsStudyGroupSurveys import QualtricsConnection
-from QualtricsStudyGroupSurveys import OathInformation
+import os
+from dotenv import load_dotenv
+from QualtricsStudyGroupSurveys import QualtricsConnection, OathInformation
+from QualtricsStudyGroupSurveys.survey_generator import generate_survey, build_people_list, get_date_choices
 
+load_dotenv()
 
 def main():
     # first command line argument is the datacenter: https://iad1.qualtrics.com/
@@ -331,4 +334,14 @@ def test_create_question(qualtrics_connection: QualtricsConnection):
 
 
 if __name__ == '__main__':
-    main()
+    start = "08-27-2027"
+    end = "09-02-2027"
+    csv_path = "data/ExampleContacts.csv"
+
+    date_choices = get_date_choices(start, end)
+    people = build_people_list(csv_path)
+    qualtrics = QualtricsConnection(os.getenv("Q_DATA_CENTER"), os.getenv("Q_API_TOKEN"))
+
+    generate_survey(qualtrics, os.getenv("Q_TEST_SURVEY_ID"), date_choices, people)
+
+    print("Done")
